Added -fno-PIC to liberty and miami builds.
By default with GCC SH, position-independent code is generated.
Supposedly this can be less-performant (although the extent to which it
is on SH4 is debatable). It also has to produce larger binaries due to
offsets in the .text segment.

Added -fno-PIC flag to miami and liberty Makefiles. Binary size dropped
by over 8KB. Performance looks the same. Don't forget to add this when
building KOS too for best results.