Moved HUD location label up to align with minimap.
Users were complaining that the location HUD strings were appearing too
low onscreen and were getting cut-off on certain CRTs and that the
placement of this text does not match that of the PS2 version and looks
janky.

This PR simply moves it up vertically to align with the bottom of the
Radar guage so that it doesn't get cut off and matches the relative
positioning of the PS2 version, which people seem to like better.