Phase 2 of liberty/miami engine+RW SH4 math.
This PR is the second batch of SH4-specific optimizations for key
linear algebra routines within both the RW layer and liberty + miami
engines.

It also includes a bunch of routines that were accelerated within
liberty but were never added to miami, which has helped contribute to
the lower perf in Miami.

1) CQuaternion
    - magnitude, squared magnitude, and dot product use SH4 instructions
      now
2) CVector (Miami port from Liberty)
    - Multiply3x3, magnitude, squared magnitude, dot product, distance
      use SH4 instructions
3) rwbase.h
    - vector3 length, vector3 dot product, and quaternion dot product
      use SH4 instructions